requirements.txt
2captcha-python
yachalk
python-dotenv
